webapp/requirements.txt
taipy
requests
Pillow
numpy